type error fix

+++ modules/task_3/narlyyev_p_gauss/main.cpp
@@ -7,8 +7,8 @@
 TEST(Parallel_Operations_MPI, Image_100x1) {
     int rank;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    int rows = 8;
-    int cols = 1;
+    size_t rows = 8;
+    size_t cols = 1;
     std::vector<double> Image;
     if (rank == 0) {
         Image = getRandomImage(rows, cols);
@@ -29,8 +29,8 @@ TEST(Parallel_Operations_MPI, Image_100x1) {
 TEST(Parallel_Operations_MPI, Image_1x100) {
     int rank;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    int rows = 1;
-    int cols = 8;
+    size_t rows = 1;
+    size_t cols = 8;
     std::vector<double> Image;
     if (rank == 0) {
         Image = getRandomImage(rows, cols);
@@ -51,8 +51,8 @@ TEST(Parallel_Operations_MPI, Image_1x100) {
 TEST(Parallel_Operations_MPI, Image_513x234) {
     int rank;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    int rows = 513;
-    int cols = 234;
+    size_t rows = 513;
+    size_t cols = 234;
     std::vector<double> Image;
     if (rank == 0) {
         Image = getRandomImage(rows, cols);
@@ -73,8 +73,8 @@ TEST(Parallel_Operations_MPI, Image_513x234) {
 TEST(Parallel_Operations_MPI, Image_234x513) {
     int rank;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    int rows = 234;
-    int cols = 512;
+    size_t rows = 234;
+    size_t cols = 512;
     std::vector<double> Image;
     if (rank == 0) {
         Image = getRandomImage(rows, cols);
@@ -95,8 +95,8 @@ TEST(Parallel_Operations_MPI, Image_234x513) {
 TEST(Parallel_Operations_MPI, Image_512x512) {
     int rank;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    int rows = 512;
-    int cols = 512;
+    size_t rows = 512;
+    size_t cols = 512;
     std::vector<double> Image;
     if (rank == 0) {
         Image = getRandomImage(rows, cols);
@@ -118,8 +118,8 @@ TEST(Parallel_Operations_MPI, Empty_Image) {
     int rank;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
     std::vector<double> Image;
-    int rows = 1;
-    int cols = 1;
+    size_t rows = 1;
+    size_t cols = 1;
 
     ASSERT_ANY_THROW(getParallelOperations(Image, rows, cols));
 }
@@ -128,8 +128,8 @@ TEST(Parallel_Operations_MPI, Invalid_Param) {
     int rank;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
     std::vector<double> Image;
-    int rows = 1;
-    int cols = 1;
+    size_t rows = 1;
+    size_t cols = 1;
     if (rank == 0) {
         Image = getRandomImage(rows, cols);
     }
